+++ CHANGELOG.md
@@ -8,6 +8,7 @@ en-US:
 - Add the feature of selecting end-of-line sequence
 - Add status bar log information, SSH ciphers information
 - Add emphasized color border to the activated session in split screen mode
+- Add custom color feature to the highlight
 - Fix the issue that clicking the new tab button in split screen mode may not create the session correctly or be located under the wrong tab group
 - Fix the issue that the SSH connection cannot be reconnected by tapping the Enter key in some cases
 
@@ -17,6 +18,7 @@ zh-CN:
 - 增加选择行尾序列功能
 - 增加状态栏日志信息，SSH加密算法信息
 - 分屏模式下激活的会话增加强调色边框
+- 高亮功能增加自定义颜色功能
 - 修复分屏模式下某些情况点击新标签按钮会话未正确创建或位于错误的标签页组下
 - 修复ssh连接部分情况下无法通过敲击回车键发起重连的问题