+++ docs/changelog.md
@@ -8,6 +8,7 @@
 - Add the feature of selecting end-of-line sequence
 - Add status bar log information, SSH ciphers information
 - Add emphasized color border to the activated session in split screen mode
+- Add custom color feature to the highlight
 - Fix the issue that clicking the new tab button in split screen mode may not create the session correctly or be located under the wrong tab group
 - Fix the issue that the SSH connection cannot be reconnected by tapping the Enter key in some cases